separate useSubscription tests
